+++ src/api/subcategory.ts
@@ -1,42 +1,54 @@
-import type { ItemType, SubcategoriesType } from '@app/types/list.types';
+'use server';
 
-export async function getSubcategoryItems(
-  id: string
-): Promise<{ subcategoryItems: ItemType[] } | undefined> {
-  try {
-    const res = await fetch(
-      `${process.env.NEXT_PUBLIC_API_URL}subcategory-items/subcategory/${id}`,
-      {
-        method: 'GET',
-        headers: {
-          Authorization: `Bearer ${process.env.NEXT_PUBLIC_TOKEN}`,
-          'Content-Type': 'application/json',
-        },
-        cache: 'no-store',
-      }
-    );
-    return res.json();
-  } catch (e) {
-    console.log('error', e);
-  }
-}
+import { revalidateTag } from 'next/cache';
+
+import type { SubcategoriesType } from '@app/types/list.types';
+
+import { getUserToken } from './user';
 
 export async function getAllSubcategories(): Promise<
   { subcategories: SubcategoriesType[] } | undefined
 > {
   try {
-    const res = await fetch(
-      `${process.env.NEXT_PUBLIC_API_URL}all-subcategories`,
-      {
-        method: 'GET',
-        headers: {
-          Authorization: `Bearer ${process.env.NEXT_PUBLIC_TOKEN}`,
-          'Content-Type': 'application/json',
-        },
-        cache: 'no-store',
-      }
-    );
-    return res.json();
+    const token = await getUserToken();
+    if (token) {
+      const res = await fetch(
+        `${process.env.NEXT_PUBLIC_API_URL}all-subcategories`,
+        {
+          method: 'GET',
+          headers: {
+            Authorization: `Bearer ${token}`,
+            'Content-Type': 'application/json',
+          },
+          next: { tags: ['subcategory'] },
+        }
+      );
+      return await res.json();
+    }
+  } catch (e) {
+    console.log('error', e);
+  }
+}
+
+export async function getSubcategory(
+  id: string
+): Promise<{ subcategory: SubcategoriesType } | undefined> {
+  try {
+    const token = await getUserToken();
+    if (token) {
+      const res = await fetch(
+        `${process.env.NEXT_PUBLIC_API_URL}subcategories/${id}`,
+        {
+          method: 'GET',
+          headers: {
+            Authorization: `Bearer ${token}`,
+            'Content-Type': 'application/json',
+          },
+          next: { tags: ['subcategory'] },
+        }
+      );
+      return await res.json();
+    }
   } catch (e) {
     console.log('error', e);
   }
@@ -47,18 +59,23 @@ export async function createSubcategory(
   data: SubcategoriesType
 ): Promise<{ subcategoryItem: SubcategoriesType } | undefined> {
   try {
-    const res = await fetch(
-      `${process.env.NEXT_PUBLIC_API_URL}subcategories/${categoryId}`,
-      {
-        method: 'POST',
-        headers: {
-          Authorization: `Bearer ${process.env.NEXT_PUBLIC_TOKEN}`,
-          'Content-Type': 'application/json',
-        },
-        body: JSON.stringify(data),
-      }
-    );
-    return res.json();
+    const token = await getUserToken();
+    if (token) {
+      const res = await fetch(
+        `${process.env.NEXT_PUBLIC_API_URL}subcategories/${categoryId}`,
+        {
+          method: 'POST',
+          headers: {
+            Authorization: `Bearer ${token}`,
+            'Content-Type': 'application/json',
+          },
+          body: JSON.stringify(data),
+        }
+      );
+      const result = await res.json();
+      revalidateTag('subcategory');
+      return result;
+    }
   } catch (e) {
     console.log('error', e);
   }
@@ -69,18 +86,23 @@ export async function updateSubcategory(
   data: SubcategoriesType
 ): Promise<{ subcategoryItem: SubcategoriesType } | undefined> {
   try {
-    const res = await fetch(
-      `${process.env.NEXT_PUBLIC_API_URL}subcategories/${categoryId}`,
-      {
-        method: 'PUT',
-        headers: {
-          Authorization: `Bearer ${process.env.NEXT_PUBLIC_TOKEN}`,
-          'Content-Type': 'application/json',
-        },
-        body: JSON.stringify(data),
-      }
-    );
-    return res.json();
+    const token = await getUserToken();
+    if (token) {
+      const res = await fetch(
+        `${process.env.NEXT_PUBLIC_API_URL}subcategories/${categoryId}`,
+        {
+          method: 'PUT',
+          headers: {
+            Authorization: `Bearer ${token}`,
+            'Content-Type': 'application/json',
+          },
+          body: JSON.stringify(data),
+        }
+      );
+      const result = await res.json();
+      revalidateTag('subcategory');
+      return result;
+    }
   } catch (e) {
     console.log('error', e);
   }
@@ -88,34 +110,17 @@ export async function updateSubcategory(
 
 export async function deleteSubcategory(id: string): Promise<void> {
   try {
-    await fetch(`${process.env.NEXT_PUBLIC_API_URL}subcategories/${id}`, {
-      method: 'DELETE',
-      headers: {
-        Authorization: `Bearer ${process.env.NEXT_PUBLIC_TOKEN}`,
-        'Content-Type': 'application/json',
-      },
-    });
-  } catch (e) {
-    console.log('error', e);
-  }
-}
-
-export async function getSubcategory(
-  id: string
-): Promise<{ subcategory: SubcategoriesType } | undefined> {
-  try {
-    const res = await fetch(
-      `${process.env.NEXT_PUBLIC_API_URL}subcategories/${id}`,
-      {
-        method: 'GET',
+    const token = await getUserToken();
+    if (token) {
+      await fetch(`${process.env.NEXT_PUBLIC_API_URL}subcategories/${id}`, {
+        method: 'DELETE',
         headers: {
-          Authorization: `Bearer ${process.env.NEXT_PUBLIC_TOKEN}`,
+          Authorization: `Bearer ${token}`,
           'Content-Type': 'application/json',
         },
-        cache: 'no-store',
-      }
-    );
-    return res.json();
+      });
+      revalidateTag('subcategory');
+    }
   } catch (e) {
     console.log('error', e);
   }
